+++ baseband/mark5b/base.py
@@ -210,16 +210,17 @@ class Mark5BStreamReader(Mark5BStreamBase, VLBIStreamReaderBase):
         squeezing). If an empty tuple (default), all channels are read.
     fill_value : float or complex
         Value to use for invalid or missing data. Default: 0.
-    verify : bool, optional
-        Whether to do basic checks of frame integrity when reading.  The first
-        frame of the stream is always checked.  Default: `True`.
+    verify : bool or 'fix', optional
+        Whether to do basic checks of frame integrity when reading.
+        Default: 'fix', which implies basic verification and replacement
+        of gaps with zeros.
     """
 
     _sample_shape_maker = Mark5BPayload._sample_shape_maker
 
     def __init__(self, fh_raw, sample_rate=None, kday=None, ref_time=None,
                  nchan=None, bps=2, squeeze=True, subset=(), fill_value=0.,
-                 verify=True):
+                 verify='fix'):
 
         if nchan is None:
             raise TypeError("Mark 5B stream reader requires nchan to be "
@@ -249,6 +250,9 @@ class Mark5BStreamReader(Mark5BStreamBase, VLBIStreamReaderBase):
         # 500 days away from the start.
         last_header.infer_kday(self.start_time)
         return last_header
+
+    def _set_time(self, header, time):
+        header.set_time(time, frame_rate=self._frame_rate)
 
     def _tell_frame(self, frame):
         # Override to provide index faster, without calculating times.
@@ -351,9 +355,10 @@ subset : indexing object, optional
     squeezing). If an empty tuple (default), all channels are read.
 fill_value : float or complex
     Value to use for invalid or missing data. Default: 0.
-verify : bool, optional
-    Whether to do basic checks of frame integrity when reading.  The first
-    frame of the stream is always checked.  Default: `True`.
+verify : bool or 'fix', optional
+    Whether to do basic checks of frame integrity when reading.
+    Default: 'fix', which implies basic verification and replacement
+    of gaps with zeros.
 
 --- For writing a stream : (see `~baseband.mark5b.base.Mark5BStreamWriter`)